60k S2 Sampler ok
