+++ egs_s2/LibriLight/local/train_60k_small.sh
@@ -12,10 +12,10 @@ omp_num=8
 # 注意 *_dirs 参数后面不可以有 ''='
 OMP_NUM_THREADS=${omp_num} python3 ${BIN_DIR}/train_librilight_60k.py \
         --config_file=${config_path} \
-        --train_semantic_dirs ''${root_dir}'/'${dump_dir}'/small/train/' ''${root_dir}'/'${dump_dir}'/medium/train/' \
-        --train_acoustic_dirs ''${root_dir}'/'${dump_dir}'/small/train/acoustic_token/'  ''${root_dir}'/'${dump_dir}'/medium/train/acoustic_token/' \
-        --dev_semantic_dirs ''${root_dir}'/'${dump_dir}'/small/dev/' ''${root_dir}'/'${dump_dir}'/medium/dev/' \
-        --dev_acoustic_dirs ''${root_dir}'/'${dump_dir}'/small/dev/acoustic_token/' ''${root_dir}'/'${dump_dir}'/medium/dev/acoustic_token/' \
+        --train_semantic_dirs ''${root_dir}'/'${dump_dir}'/small/train/' \
+        --train_acoustic_dirs ''${root_dir}'/'${dump_dir}'/small/train/acoustic_token/' \
+        --dev_semantic_dirs ''${root_dir}'/'${dump_dir}'/small/dev/' \
+        --dev_acoustic_dirs ''${root_dir}'/'${dump_dir}'/small/dev/acoustic_token/' \
         --output=${root_dir}/${train_output_path} \
         --log_frequency=${log_frequency} \
         --dist_url=${dist_url} \
